Refactor packages
Refactor classes into different packages for better separation of concerns.
In the future, a module-info.java can be added more easily now, however, JPMS create difficulties with tests, since they would need to be in their own project due to test-module dependencies being different than application-module dependencies and JPMS not supporting this separation.
