+++ src/main/java/ca/trovo/greetings/Greeter.java
@@ -1,4 +1,4 @@
-package ca.trovo.hello;
+package ca.trovo.greetings;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
