+++ src/main/java/ca/trovo/greetings/greeter/RandomListGreeter.java
@@ -1,10 +1,10 @@
-package ca.trovo.hello.greeter;
+package ca.trovo.greetings.greeter;
 
 import java.util.List;
 import java.util.Random;
 
-import ca.trovo.hello.Greeter;
-import ca.trovo.hello.Greeting;
+import ca.trovo.greetings.Greeter;
+import ca.trovo.greetings.Greeting;
 
 
 /**
