+++ src/test/java/ca/trovo/BaseIntegrationTestSuite.java
@@ -1,8 +1,9 @@
-package ca.trovo.hello;
+package ca.trovo;
 
 import java.io.File;
+import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
+import java.util.stream.Collectors;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.Archive;
@@ -11,7 +12,7 @@ import org.jboss.shrinkwrap.api.Archive;
 /**
  * Base class for all integration tests for this project that are going to run in {@code Arquillian}.
  */
-abstract class BaseIntegrationTestSuite extends ArquillianDeploymentBuilder {
+public abstract class BaseIntegrationTestSuite extends ArquillianDeploymentBuilder {
 
 	/** The {@code .war} file name that is the output of this project. Defined/Created by {@code Maven pom.xml}. */
 	private static final String DEPLOYMENT_FILE_NAME = "hello-jakartaee.war";
@@ -23,7 +24,7 @@ abstract class BaseIntegrationTestSuite extends ArquillianDeploymentBuilder {
 	 *     <li>The real application's {@link #DEPLOYMENT_FILE_NAME} into its container (e.g. {@code Wildfly})</li>
 	 *     <li>Add, build and deploy the test classes defined in this package as part of the deployment.</li>
 	 * </ol>
-	 * This will include all classes under this package ("ca.trovo.hello.test")
+	 * This will include all classes under this package
 	 *
 	 * @return The {@code Arquillian} archive for deployment.
 	 *
@@ -31,7 +32,9 @@ abstract class BaseIntegrationTestSuite extends ArquillianDeploymentBuilder {
 	 */
 	@Deployment
     public static Archive<?> buildTestWar() {
-		return buildTestWar(List.of(BaseIntegrationTestSuite.class.getPackageName()));
+		return buildTestWar(
+			findPackageNamesStartingWith(BaseIntegrationTestSuite.class.getPackageName())
+		);
 	}
 
 
@@ -69,5 +72,22 @@ abstract class BaseIntegrationTestSuite extends ArquillianDeploymentBuilder {
 	 */
 	public static Archive<?> buildTestWar(Collection<String> packages, File exportedTestWarOutput) {
 		return ArquillianDeploymentBuilder.buildTestWar(ArquillianDeploymentBuilder.DEPLOYMENT_FILE_PATH, DEPLOYMENT_FILE_NAME, packages, exportedTestWarOutput);
+	}
+
+
+	/**
+	 * Helper method to find all packages in the classpath where the name starts with the given {@code prefix}.
+	 *
+	 * @param prefix
+	 * 			The package's prefix to filter.
+	 *
+	 * @return The name of the packages in the classpath where it starts with the given {@code prefix}.
+	 *
+	 */
+	public static Collection<String> findPackageNamesStartingWith(String prefix) {
+	    return Arrays.stream(Package.getPackages())
+	        .map(Package::getName)
+	        .filter(name -> name.startsWith(prefix))
+	        .collect(Collectors.toList());
 	}
 }
