+++ src/test/java/ca/trovo/greetings/greeter/GreeterFactoryTestHelper.java
@@ -1,9 +1,10 @@
-package ca.trovo.hello.greeter;
+package ca.trovo.greetings.greeter;
 
 import java.util.Collection;
 import java.util.Map;
 
-import ca.trovo.hello.Greeter;
+import ca.trovo.greetings.Greeter;
+import ca.trovo.greetings.GreeterFactory;
 
 
 /**
