+++ src/test/java/ca/trovo/hello/HelloResouceIT.java
@@ -10,6 +10,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import ca.trovo.BaseIntegrationTestSuiteHttpClient;
 import ca.trovo.utils.HttpUtils;
 import jakarta.ws.rs.core.Response;
